moved vcf related code to vcf.py module
